pyproject.toml
[project]
name = "mongodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
]
